"added UnhealthyToHealthyCompositeKey and altered equals and hashcode for FoodToBrandCompositeKey"
